move physical sort
