+++ datafusion/physical-expr-core/Cargo.toml
@@ -32,18 +32,6 @@ rust-version = { workspace = true }
 name = "datafusion_physical_expr_core"
 path = "src/lib.rs"
 
-# TODO: Remove unused features
-
-[features]
-crypto_expressions = ["md-5", "sha2", "blake2", "blake3"]
-default = [
-    "crypto_expressions",
-    "regex_expressions",
-    "encoding_expressions",
-]
-encoding_expressions = ["base64", "hex"]
-regex_expressions = ["regex"]
-
 [dependencies]
 ahash = { version = "0.8", default-features = false, features = [
     "runtime-rng",
